Update README.md
Hi Rishi,

I have edited the original content directly on the README.md file on GitHub.  It tells me that my edits will be in a new fork so I'm sure you will be able to accept and import it as applicable.  Please let me know if you have any questions.    Thanks.  -shoba